+++ src/animations/ALeft.hpp
@@ -14,9 +14,9 @@ private:
 
 public:
 
-  ALeft(Character * _parent,const sf::Int32 _moduloTime);
+  ALeft(Character * _parent,const sf::Int64 _moduloTime);
 
-  void execute(const sf::Int32 elapsed);
+  void execute(const sf::Int64 elapsed);
 };
 
 #endif
