+++ src/animations/ARight.cpp
@@ -1,10 +1,10 @@
 #include "ARight.hpp"
 
 ARight::ARight(Character * _parent,
-	     const sf::Int32 _moduloTime) : AAnimation(_parent,_moduloTime) {
+	     const sf::Int64 _moduloTime) : AAnimation(_parent,_moduloTime) {
 }
 
-void ARight::execute(const sf::Int32 elapsed) {
+void ARight::execute(const sf::Int64 elapsed) {
   // TODO: Verif' they are existing components
   // #1 get all components
   CAcceleration * acceleration;
