+++ src/animations/ARight.hpp
@@ -15,9 +15,9 @@ private:
 
 public:
 
-  ARight(Character * _parent,const sf::Int32 _moduloTime);
+  ARight(Character * _parent,const sf::Int64 _moduloTime);
 
-  void execute(const sf::Int32 elapsed);
+  void execute(const sf::Int64 elapsed);
 };
 
 #endif
